- Added exception 404 with a custom message - Changed project structure

+++ src/main/java/com/litequizapp/service/CategoryRestService.java
@@ -1,5 +1,8 @@
-package com.litequizapp.server;
+package com.litequizapp.service;
 
+import com.litequizapp.entity.CategoryEntity;
+import com.litequizapp.exception.ElementNotFoundException;
+import com.litequizapp.repository.CategoryRepository;
 import java.util.ArrayList;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,6 +22,9 @@ public class CategoryRestService {
 
   public String getCategoryById(long id) {
     CategoryEntity category = categoryRepository.findById(id);
+    if (category == null) {
+      throw new ElementNotFoundException();
+    }
     return category.toString();
 
   }
@@ -34,15 +40,23 @@ public class CategoryRestService {
 
   public void createCategory(String title) {
     categoryRepository.save(new CategoryEntity(title));
+
   }
 
   public void updateCategory(long id, String title) {
     CategoryEntity category = categoryRepository.findById(id);
+    if (category == null) {
+      throw new ElementNotFoundException();
+    }
     category.setTitle(title);
     categoryRepository.save(category);
   }
 
   public void deleteCategory(long id) {
+    CategoryEntity category = categoryRepository.findById(id);
+    if (category == null) {
+      throw new ElementNotFoundException();
+    }
     categoryRepository.deleteById(id);
 
   }
